+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/dialogs/SequenceDialog.tsx
@@ -1,13 +1,5 @@
 import React, { Suspense, useRef, useState } from 'react'
-import {
-  Button,
-  DialogContent,
-  DialogActions,
-  FormControl,
-  MenuItem,
-  Select,
-  Typography,
-} from '@mui/material'
+import { Button, DialogContent, DialogActions, Typography } from '@mui/material'
 import { Dialog, ErrorMessage, LoadingEllipses } from '@jbrowse/core/ui'
 import { makeStyles } from 'tss-react/mui'
 import { observer } from 'mobx-react'
@@ -18,6 +10,7 @@ import { SimpleFeatureSerialized } from '../../../util'
 import { BaseFeatureWidgetModel } from '../../stateModelFactory'
 import SequencePanel from '../SequencePanel'
 import SequenceFeatureMenu from './SequenceFeatureMenu'
+import SequenceTypeSelector from './SequenceTypeSelector'
 
 const useStyles = makeStyles()({
   dialogContent: {
@@ -39,14 +32,10 @@ const SequenceDialog = observer(function ({
   model: BaseFeatureWidgetModel
 }) {
   const { sequenceFeatureDetails } = model
-  const { intronBp, upDownBp } = sequenceFeatureDetails
+  const { upDownBp } = sequenceFeatureDetails
   const { classes } = useStyles()
   const seqPanelRef = useRef<HTMLDivElement>(null)
-
   const [force, setForce] = useState(false)
-  const hasCDS = feature.subfeatures?.some(sub => sub.type === 'CDS')
-  const hasExon = feature.subfeatures?.some(sub => sub.type === 'exon')
-  const hasExonOrCDS = hasExon || hasCDS
   const { sequence, error } = useFeatureSequence(
     model,
     feature,
@@ -54,9 +43,7 @@ const SequenceDialog = observer(function ({
     force,
   )
 
-  const [mode, setMode] = useState(
-    hasCDS ? 'cds' : hasExon ? 'cdna' : 'genomic',
-  )
+  const [mode, setMode] = useState('cds')
   return (
     <Dialog
       maxWidth="xl"
@@ -66,67 +53,14 @@ const SequenceDialog = observer(function ({
     >
       <DialogContent className={classes.dialogContent}>
         <div>
-          <FormControl className={classes.formControl}>
-            <Select
-              size="small"
-              value={mode}
-              onChange={event => setMode(event.target.value)}
-            >
-              {Object.entries({
-                ...(hasCDS
-                  ? {
-                      cds: 'CDS',
-                    }
-                  : {}),
-                ...(hasCDS
-                  ? {
-                      protein: 'Protein',
-                    }
-                  : {}),
-                ...(hasExonOrCDS
-                  ? {
-                      cdna: 'cDNA',
-                    }
-                  : {}),
-                ...(hasExonOrCDS
-                  ? {
-                      gene: `Genomic w/ full introns`,
-                    }
-                  : {}),
-                ...(hasExonOrCDS
-                  ? {
-                      gene_updownstream: `Genomic w/ full introns +/- ${upDownBp}bp up+down stream`,
-                    }
-                  : {}),
-                ...(hasExonOrCDS
-                  ? {
-                      gene_collapsed_intron: `Genomic w/ ${intronBp}bp intron`,
-                    }
-                  : {}),
-                ...(hasExonOrCDS
-                  ? {
-                      gene_updownstream_collapsed_intron: `Genomic w/ ${intronBp}bp intron +/- ${upDownBp}bp up+down stream `,
-                    }
-                  : {}),
-
-                ...(!hasExonOrCDS
-                  ? {
-                      genomic: 'Genomic',
-                    }
-                  : {}),
-                ...(!hasExonOrCDS
-                  ? {
-                      genomic_sequence_updownstream: `Genomic +/- ${upDownBp}bp up+down stream`,
-                    }
-                  : {}),
-              }).map(([key, val]) => (
-                <MenuItem key={key} value={key}>
-                  {val}
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
+          <SequenceTypeSelector
+            mode={mode}
+            setMode={setMode}
+            feature={feature}
+            model={sequenceFeatureDetails}
+          />
           <SequenceFeatureMenu
+            mode={mode}
             ref={seqPanelRef}
             model={sequenceFeatureDetails}
           />
